Creates a draft common interface for the input
The idea behind this is that each platform will have it's own way of
representing specific inputs, so we need to have a common way of
representing the input. I decided on using a uint8 since this should be
common to all platforms.

For now, the only defined inputs are left, right, open (to open a cell)
and a generic ignored value, that we can use when any other key is
pressed.