Ncurses input (#9)
* Reads user input in a loop

The idea is that user input will be read, some action taken by the game,
then the loop continues until gameover.

For the ncurses platform, we rely on the fact that ncurses' `initscr`
sets up a SIGINT handler that attempts to clean the screen[1]

[1]: https://invisible-island.net/ncurses/man/curs_initscr.3x.html#h3-Signal-Handlers

* Stop echoing input characters in ncurses library

* Creates a draft common interface for the input

The idea behind this is that each platform will have it's own way of
representing specific inputs, so we need to have a common way of
representing the input. I decided on using a uint8 since this should be
common to all platforms.

For now, the only defined inputs are left, right, open (to open a cell)
and a generic ignored value, that we can use when any other key is
pressed.

* Display closed cells with # in ncurses

* Fixes duplicated input handler

* Enables Keypad[1] in ncurses to interpret for arrow keys correctly

[1]: https://pubs.opengroup.org/onlinepubs/7908799/xcurses/keypad.html

* Redraws minefield in the main loop

We need to redraw the screen after we handle the input.

* Tracks current cell and highlights it in ncurses

* Uses '\n' for ENTER key

KEY_ENTER in ncurses is only the keypad enter.

* [ncurses] Always stop highlighting cell in the end of the loop

* Implement handlers for left, right and open inputs

* [ncurses] Prints empty string to empty cell

This fixes a bug where cells with no bombs and no neighbors would retain
the # placeholder from when they were closed.