[MS-DOS] Compress tileset with custom compression scheme
This uses a trivial RLE scheme to compress the image, and encode it
using 2 pixels per byte (as we only use 15 colors).

File size difference:

	Original mines.xpm:   4744 bytes
	Compressed mines.bin: 1186 bytes

Almost a 4x decrease, which isn't bad for something this trivial.

Additionally, also encode the tile offsets in an uint8_t instead of
an uint16_t, saving further 100 bytes in the final executable.